pyproject.toml
[tool.poetry]
name = "cavcrosby-autotag"
version = "1.0.0"
description = "Framework for automating the creation of git tags"
authors = ["Conner Crosby <conner@cavcrosby.tech>"]
license = "MIT"
packages = [
    { include = "autotag" },
]

[tool.poetry.dependencies]
python = "^3.9.5"
pylib-cavcrosby = {git = "https://github.com/cavcrosby/pylib.git", branch = "main"}

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
flake8-docstrings = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
